PR review

+++ substrate/frame/revive/src/vm/evm/memory.rs
@@ -58,7 +58,7 @@ impl Memory {
 	pub fn resize(&mut self, offset: usize, len: usize) -> ControlFlow<Halt> {
 		let current_len = self.0.len();
 		let target_len = revm::interpreter::num_words(offset.saturating_add(len)) * 32;
-		if target_len as u32 > crate::limits::code::BASELINE_MEMORY_LIMIT {
+		if target_len > crate::limits::code::BASELINE_MEMORY_LIMIT as usize {
 			log::debug!(target: crate::LOG_TARGET, "check memory bounds failed: offset={offset} target_len={target_len} current_len={current_len}");
 			return ControlFlow::Break(HaltReason::MemoryOOG.into());
 		}
